Seperated versionCheck and some dictionaries from main.py
